Create README.md
# Interactive-Food-Pyramid

Assignment	1	-	An	Interactive	Food	Pyramid.
For	this	assignment,	you	are	required	to	develop	an	interactive	Food	Pyramid	chart	using	HTML/CSS/
The	Food	Pyramid	shows	how	much	of	what	you	eat	overall	should	come	from	each	shelf	to	achieve	
a	healthy,	balanced	diet.	The	shape	of	the	Food	Pyramid	shows	the	types	of	foods	and	drinks	people	
need	 to	 eat	most	 for	 healthy	 eating. It	 is	 divided	 into	 six	 shelves	 and	 each	 provides
you	 with	 the range	of	nutrients	and	energy	needed	for	good	health.	Healthy	eating	is	all	about	choosing	
the	right	amounts	from	each	shelf. An	interactive	Food	Pyramid	allows	you	to	record	the	number	of	foods	consumed
in	any	given	day	by categorising	and	recording	the	food	types	for	each	shelf.	

Assignment	01	-	Requirements
Your	assignment	is	to	create	an	Interactive	food	diary	that	allows	users	to	record	the	different	food	
shelf	details	for	today,	and	to	record	today!s	date.	For	this	assignment,	you	do	not	need	to	create	a
database,	etc.	just	work	on	the	"front	end”,	i.e.	the	food	pyramid	and	interactive	elements.

You	must	use	HTML	elements	to	create	the	chart	-	you	may	not	use	images	to	create	the	structure	of	
the	individual	components	(trapezoids	and	triangle)	layered	as	a	pyramid,	i.e.	you	cannot	just	create	
an	image,	load	it	and	overlay	it	with	 the	interactive	components.	Colour	and	other	 styling	 settings	
must	be	achieved	using	CSS.	You	must	use	"pure”	Javascript	(JS)	only.	Do	not	use	jQuery,	or	any	Javascript	
Framework.	Figure	1	is	a	 template	 -	 feel	 free	 to	style	 the	layers	or	buttons	according	 to	your	
preferences.	Just	make	sure	you	meet	the	fundamental	requirements. Your	food	pyramid	must	have	the	following	
interactive	functionality: It	should	provide	an	indicator	showing	 the	number	of	items	consumed	 for	a	particular
shelf,	 for	example,	in	 the	figure	 shown,	it shows	6	items	on	the	"Vegetables,	salad	and	fruit”	shelf	(the	bottom	one).	
It	should	also	include	buttons	(or	similar	devices)	to	increase	and	decrease	the	number	of	items	on	the	shelf.		
The	(+	/	-)	buttons	(controls)	should	only	appear	when	you	hover	over	a	particular	shelf.	The	shelf	numbers	(6	for	
the	lowest	shelf)	should	always	be	displayed.	You	should	also	have	some	interactive	method	for	inputting	the	date	
for	the	daily	record.	You	do	not	need	to	validate	the	date.